feat(live-state): organization procedures and disabled collection update
Plan §11: extract organization route with withProcedures (create, update,
API key helpers); block generic insert/update; migrate web and digest
worker to mutate.organization.update({ id, ... }); add optimistic org patch.

Made-with: Cursor
