+++ apps/api/src/live-state/router.ts
@@ -1,10 +1,6 @@
 import { router as createRouter } from "@live-state/sync/server";
-import { addYears } from "date-fns";
 import { ulid } from "ulid";
 import { z } from "zod";
-import { publicKeys } from "../lib/api-key";
-import { dodopayments } from "../lib/payment";
-import { sendWelcomeEmail } from "../trigger/send-welcome-email";
 import { privateRoute, publicRoute } from "./factories";
 import { agentChatMessageRoute, agentChatRoute } from "./router/agent-chat";
 import { allowlistRoute } from "./router/allowlist";
@@ -14,6 +10,7 @@ import { inviteRoute } from "./router/invite";
 import labelsRoute from "./router/labels";
 import messageRoute from "./router/message";
 import onboardingRoute from "./router/onboarding";
+import { organizationRoute } from "./router/organization";
 import { organizationUserRoute } from "./router/organization-user";
 import { slackChannelsCache } from "./router/slack-channels";
 import { subscriptionRoute } from "./router/subscription";
@@ -40,270 +37,7 @@ const isPostgresUniqueViolation = (error: unknown): boolean => {
 export const router = createRouter({
   schema,
   routes: {
-    organization: publicRoute
-      .collectionRoute(schema.organization, {
-        read: () => true,
-        insert: () => false,
-        update: {
-          preMutation: ({ ctx }) => {
-            if (ctx?.internalApiKey) return true;
-            if (!ctx?.session) return false;
-
-            return {
-              organizationUsers: {
-                userId: ctx.session.userId,
-                role: "owner",
-                enabled: true,
-              },
-            };
-          },
-          postMutation: ({ ctx }) => {
-            if (ctx?.internalApiKey) return true;
-            if (!ctx?.session) return false;
-
-            return {
-              organizationUsers: {
-                userId: ctx.session.userId,
-                role: "owner",
-                enabled: true,
-              },
-            };
-          },
-        },
-      })
-      .withMutations(({ mutation }) => ({
-        create: mutation(
-          z.object({
-            name: z.string(),
-            slug: z
-              .string()
-              .min(4)
-              .refine(
-                (slug) => {
-                  // TODO: Unify reserved slugs list - extract to shared constant
-                  const reservedSlugs = [
-                    "support",
-                    "help",
-                    "status",
-                    "api",
-                    "admin",
-                    "www",
-                    "app",
-                    "dashboard",
-                    "login",
-                    "signup",
-                    "register",
-                    "account",
-                    "settings",
-                    "billing",
-                    "docs",
-                    "documentation",
-                    "blog",
-                    "about",
-                    "contact",
-                    "privacy",
-                    "terms",
-                    "legal",
-                  ];
-                  return !reservedSlugs.includes(slug.toLowerCase());
-                },
-                {
-                  message: "This slug is reserved and cannot be used",
-                },
-              ),
-          }),
-        ).handler(async ({ req, db }) => {
-          const organizationId = ulid().toLowerCase();
-
-          const dodopaymentsCustomer = await dodopayments?.customers.create({
-            email: req.context.user?.email,
-            name: req.context.user?.name,
-          });
-
-          const organization = await db.organization.insert({
-            id: organizationId,
-            name: req.input!.name,
-            slug: req.input!.slug,
-            createdAt: new Date(),
-            logoUrl: null,
-            socials: null,
-            customInstructions: null,
-            settings: null,
-          });
-
-          await db.subscription.insert({
-            id: ulid().toLowerCase(),
-            organizationId,
-            customerId: dodopaymentsCustomer?.customer_id ?? null,
-            subscriptionId: null,
-            plan: "trial",
-            status: null,
-            seats: 1,
-            createdAt: new Date(),
-            updatedAt: new Date(),
-          });
-
-          await db.organizationUser.insert({
-            id: ulid().toLowerCase(),
-            organizationId,
-            userId: req.context.session.userId,
-            enabled: true,
-            role: "owner",
-          });
-
-          // Send welcome email if this is the user's first organization
-          const userMemberships = Object.values(
-            await db.find(schema.organizationUser, {
-              where: {
-                userId: req.context.session.userId,
-              },
-            }),
-          );
-
-          if (userMemberships.length === 1) {
-            const delayMinutes = Math.floor(Math.random() * 21) + 20;
-
-            sendWelcomeEmail
-              .trigger(
-                {
-                  email: req.context.user!.email,
-                  name: req.context.user!.name,
-                },
-                { delay: `${delayMinutes}m` },
-              )
-              .catch((err) => {
-                console.error("Failed to schedule welcome email", err);
-              });
-          }
-
-          return {
-            success: true,
-            organization,
-          };
-        }),
-        createPublicApiKey: mutation(
-          z.object({
-            organizationId: z.string(),
-            expiresAt: z.iso.datetime().optional(),
-            name: z.string().optional(),
-          }),
-        ).handler(async ({ req, db }) => {
-          const organizationId = req.input.organizationId;
-
-          let authorized = !!req.context?.internalApiKey;
-
-          if (!authorized && req.context?.session?.userId) {
-            const selfOrgUser = Object.values(
-              await db.find(schema.organizationUser, {
-                where: {
-                  organizationId,
-                  userId: req.context.session.userId,
-                },
-                include: {
-                  user: true,
-                  organization: true,
-                },
-              }),
-            )[0] as any;
-
-            authorized = selfOrgUser && selfOrgUser.role === "owner";
-          }
-
-          if (!authorized) {
-            throw new Error("UNAUTHORIZED");
-          }
-
-          const publicApiKey = await publicKeys.create({
-            ownerId: organizationId,
-            tags: ["organization"],
-            expiresAt:
-              req.input.expiresAt ?? addYears(new Date(), 1).toISOString(),
-            name: req.input.name,
-          });
-
-          return {
-            id: publicApiKey.record.id,
-            key: publicApiKey.key,
-            expiresAt: publicApiKey.record.metadata.expiresAt,
-            name: publicApiKey.record.metadata.name,
-          };
-        }),
-        revokePublicApiKey: mutation(
-          z.object({
-            id: z.string(),
-          }),
-        ).handler(async ({ req, db }) => {
-          if (!req.context?.session?.userId) {
-            throw new Error("UNAUTHORIZED");
-          }
-
-          const publicApiKey = await publicKeys.findById(req.input.id);
-
-          if (!publicApiKey) {
-            throw new Error("PUBLIC_API_KEY_NOT_FOUND");
-          }
-
-          const selfOrgUser = Object.values(
-            await db.find(schema.organizationUser, {
-              where: {
-                organizationId: publicApiKey.metadata.ownerId,
-                userId: req.context.session.userId,
-              },
-            }),
-          )[0] as any;
-
-          if (!selfOrgUser || selfOrgUser.role !== "owner") {
-            throw new Error("UNAUTHORIZED");
-          }
-
-          await publicKeys.revoke(publicApiKey.id).catch((error) => {
-            console.error("Error revoking public API key", error);
-            throw new Error("FAILED_TO_REVOKE_PUBLIC_API_KEY");
-          });
-
-          return {
-            success: true,
-          };
-        }),
-        listApiKeys: mutation(
-          z.object({
-            organizationId: z.string(),
-          }),
-        ).handler(async ({ req, db }) => {
-          const organizationId = req.input.organizationId;
-
-          let authorized = !!req.context?.internalApiKey;
-
-          if (!authorized && req.context?.session?.userId) {
-            const selfOrgUser = Object.values(
-              await db.find(schema.organizationUser, {
-                where: {
-                  organizationId,
-                  userId: req.context.session.userId,
-                },
-              }),
-            )[0] as any;
-
-            authorized = selfOrgUser && selfOrgUser.role === "owner";
-          }
-
-          if (!authorized) {
-            throw new Error("UNAUTHORIZED");
-          }
-
-          const apiKeys = await publicKeys.list(organizationId);
-
-          return apiKeys
-            .filter((apiKey) => !apiKey.metadata.revokedAt)
-            .map((apiKey) => ({
-              id: apiKey.id,
-              expiresAt: apiKey.metadata.expiresAt,
-              name: apiKey.metadata.name,
-              type: "public",
-              createdAt: apiKey.metadata.createdAt,
-            }));
-        }),
-      })),
+    organization: organizationRoute,
     organizationUser: organizationUserRoute,
     user: userRoute,
     author: authorRoute,
