+++ apps/web/src/routes/app/_workspace/settings/organization/index.tsx
@@ -212,7 +212,8 @@ function OrgProfileForm({ org, currentOrg, isUserOwner }: OrgFormProps) {
         timezone: value.orgTimezone,
       });
 
-      mutate.organization.update(currentOrg.id, {
+      mutate.organization.update({
+        id: currentOrg.id,
         name: value.orgName,
         slug: value.orgSlug,
         logoUrl,
@@ -447,7 +448,8 @@ function DigestSettingsForm({ org, currentOrg, isUserOwner }: OrgFormProps) {
         },
       });
 
-      mutate.organization.update(currentOrg.id, {
+      mutate.organization.update({
+        id: currentOrg.id,
         settings: nextSettings,
       });
     },
